+++ components/sup/src/ctl_gateway/server.rs
@@ -44,7 +44,7 @@ use tokio_codec::Framed;
 use tokio_core::reactor;
 
 use super::{CtlRequest, REQ_TIMEOUT};
-use manager::{Manager, ManagerState};
+use manager::{commands, ManagerState};
 
 /// Sending half of an mpsc unbounded channel used for sending replies for a transactional message
 /// from the main thread back to the CtlGateway. This half is stored in a
@@ -287,7 +287,7 @@ impl Future for SrvHandler {
                                 CtlCommand::new(
                                     Some(self.tx.clone()),
                                     msg.transaction(),
-                                    move |state, req| Manager::service_cfg(state, req, m.clone()),
+                                    move |state, req| commands::service_cfg(state, req, m.clone()),
                                 )
                             }
                             "SvcFilePut" => {
@@ -298,7 +298,7 @@ impl Future for SrvHandler {
                                     Some(self.tx.clone()),
                                     msg.transaction(),
                                     move |state, req| {
-                                        Manager::service_file_put(state, req, m.clone())
+                                        commands::service_file_put(state, req, m.clone())
                                     },
                                 )
                             }
@@ -310,7 +310,7 @@ impl Future for SrvHandler {
                                     Some(self.tx.clone()),
                                     msg.transaction(),
                                     move |state, req| {
-                                        Manager::service_cfg_set(state, req, m.clone())
+                                        commands::service_cfg_set(state, req, m.clone())
                                     },
                                 )
                             }
@@ -322,7 +322,7 @@ impl Future for SrvHandler {
                                     Some(self.tx.clone()),
                                     msg.transaction(),
                                     move |state, req| {
-                                        Manager::service_cfg_validate(state, req, m.clone())
+                                        commands::service_cfg_validate(state, req, m.clone())
                                     },
                                 )
                             }
@@ -333,7 +333,7 @@ impl Future for SrvHandler {
                                 CtlCommand::new(
                                     Some(self.tx.clone()),
                                     msg.transaction(),
-                                    move |state, req| Manager::service_load(state, req, m.clone()),
+                                    move |state, req| commands::service_load(state, req, m.clone()),
                                 )
                             }
                             "SvcUnload" => {
@@ -344,7 +344,7 @@ impl Future for SrvHandler {
                                     Some(self.tx.clone()),
                                     msg.transaction(),
                                     move |state, req| {
-                                        Manager::service_unload(state, req, m.clone())
+                                        commands::service_unload(state, req, m.clone())
                                     },
                                 )
                             }
@@ -355,7 +355,9 @@ impl Future for SrvHandler {
                                 CtlCommand::new(
                                     Some(self.tx.clone()),
                                     msg.transaction(),
-                                    move |state, req| Manager::service_start(state, req, m.clone()),
+                                    move |state, req| {
+                                        commands::service_start(state, req, m.clone())
+                                    },
                                 )
                             }
                             "SvcStop" => {
@@ -365,7 +367,7 @@ impl Future for SrvHandler {
                                 CtlCommand::new(
                                     Some(self.tx.clone()),
                                     msg.transaction(),
-                                    move |state, req| Manager::service_stop(state, req, m.clone()),
+                                    move |state, req| commands::service_stop(state, req, m.clone()),
                                 )
                             }
                             "SvcStatus" => {
@@ -376,7 +378,7 @@ impl Future for SrvHandler {
                                     Some(self.tx.clone()),
                                     msg.transaction(),
                                     move |state, req| {
-                                        Manager::service_status(state, req, m.clone())
+                                        commands::service_status(state, req, m.clone())
                                     },
                                 )
                             }
@@ -388,7 +390,7 @@ impl Future for SrvHandler {
                                     Some(self.tx.clone()),
                                     msg.transaction(),
                                     move |state, req| {
-                                        Manager::supervisor_depart(state, req, m.clone())
+                                        commands::supervisor_depart(state, req, m.clone())
                                     },
                                 )
                             }
